Don't include doctests for now.
The RST conversion was failing because the ./tree/master/src/... path
wasn't resolving. Changing that to ./src/collective/... works, but
then listingviews.rst causes a bunch of RST errors. The doctests would
have to be edited.

Similarly, hyperlinking to a relative path won't work. Do we want to
link to the file in github instead?